pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "unity-mcp"
version = "0.1.0"
description = "Unity Model Context Protocol Server Implementation"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "pydantic>=2.0.0",
    "websockets>=11.0.0",
    "pillow>=10.0.0",  # For image processing
    "numpy>=1.20.0",   # For numerical operations
    "mcp>=0.1.0",      # Model Context Protocol
    "httpx>=0.24.0",   # For HTTP requests
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "httpx>=0.28.0",
]

[tool.setuptools]
packages = ["server"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.mypy]
python_version = "3.8"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"